requirements.txt
python-telegram-bot==21.0
groq
python-dotenv